requirements.txt
alembic==1.5.8
aniso8601==9.0.1
autopep8==1.5.7
boto3==1.18.57
botocore==1.21.57
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
importlib-metadata==3.7.3
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
python-dateutil==2.8.1
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
s3transfer==0.5.0
six==1.15.0
SQLAlchemy==1.4.3
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==1.0.1
zipp==3.4.1
